Split make request api and output
