+++ .gitignore
@@ -25,3 +25,5 @@ go.work.sum
 .env
 
 bin/
+
+test-api/
