+++ neo4j/aura/internal/subcommands/customermanagedkey/get.go
@@ -2,8 +2,10 @@ package customermanagedkey
 
 import (
 	"fmt"
+	"net/http"
 
 	"github.com/neo4j/cli/neo4j/aura/internal/api"
+	"github.com/neo4j/cli/neo4j/aura/internal/output"
 	"github.com/spf13/cobra"
 )
 
@@ -14,7 +16,18 @@ func NewGetCmd() *cobra.Command {
 		Long:  `This subcommand returns details about a specific Customer Managed Key.`,
 		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return api.MakeRequest(cmd, "GET", fmt.Sprintf("/customer-managed-keys/%s", args[0]), nil)
+			path := fmt.Sprintf("/customer-managed-keys/%s", args[0])
+			resBody, statusCode, err := api.MakeRequest(cmd, http.MethodGet, path, nil)
+			if err != nil {
+				return err
+			}
+
+			if statusCode == http.StatusOK {
+				output.PrintBody(cmd, resBody)
+
+			}
+
+			return nil
 		},
 	}
 }
