+++ neo4j/aura/internal/subcommands/instance/pause.go
@@ -2,8 +2,10 @@ package instance
 
 import (
 	"fmt"
+	"net/http"
 
 	"github.com/neo4j/cli/neo4j/aura/internal/api"
+	"github.com/neo4j/cli/neo4j/aura/internal/output"
 	"github.com/spf13/cobra"
 )
 
@@ -20,7 +22,22 @@ The pause time depends on the amount of data stored in the instance; larger quan
 If another operation is being performed on the instance you are trying to pause, an error will be returned that indicates that the pause operation cannot be performed.`,
 		Args: cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return api.MakeRequest(cmd, "POST", fmt.Sprintf("/instances/%s/pause", args[0]), nil)
+			path := fmt.Sprintf("/instances/%s/pause", args[0])
+
+			resBody, statusCode, err := api.MakeRequest(cmd, http.MethodPost, path, nil)
+			if err != nil {
+				return err
+			}
+
+			// NOTE: Instance pause should not return OK (200), it always returns 202
+			if statusCode == http.StatusAccepted || statusCode == http.StatusOK {
+				err = output.PrintBody(cmd, resBody)
+				if err != nil {
+					return err
+				}
+
+			}
+			return nil
 		},
 	}
 }
