+++ neo4j/aura/internal/subcommands/instance/resume.go
@@ -5,6 +5,7 @@ import (
 	"net/http"
 
 	"github.com/neo4j/cli/neo4j/aura/internal/api"
+	"github.com/neo4j/cli/neo4j/aura/internal/output"
 	"github.com/spf13/cobra"
 )
 
@@ -19,7 +20,22 @@ Resuming an instance is an asynchronous operation. You can poll the current stat
 If another operation is being performed on the instance you are trying to resume, an error will be returned that indicates that resume cannot be performed.`,
 		Args: cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return api.MakeRequest(cmd, http.MethodPost, fmt.Sprintf("/instances/%s/resume", args[0]), nil)
+			path := fmt.Sprintf("/instances/%s/resume", args[0])
+
+			resBody, statusCode, err := api.MakeRequest(cmd, http.MethodPost, path, nil)
+			if err != nil {
+				return err
+			}
+
+			// NOTE: Instance resume should not return OK (200), it always returns 202
+			if statusCode == http.StatusAccepted || statusCode == http.StatusOK {
+				err = output.PrintBody(cmd, resBody)
+				if err != nil {
+					return err
+				}
+
+			}
+			return nil
 		},
 	}
 }
