+++ neo4j/aura/internal/subcommands/tenant/get.go
@@ -2,8 +2,10 @@ package tenant
 
 import (
 	"fmt"
+	"net/http"
 
 	"github.com/neo4j/cli/neo4j/aura/internal/api"
+	"github.com/neo4j/cli/neo4j/aura/internal/output"
 	"github.com/spf13/cobra"
 )
 
@@ -14,7 +16,22 @@ func NewGetCmd() *cobra.Command {
 		Long:  "This subcommand returns details about a specific Aura Tenant.",
 		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return api.MakeRequest(cmd, "GET", fmt.Sprintf("/tenants/%s", args[0]), nil)
+			path := fmt.Sprintf("/tenants/%s", args[0])
+
+			resBody, statusCode, err := api.MakeRequest(cmd, http.MethodGet, path, nil)
+			if err != nil {
+				return err
+			}
+
+			if statusCode == http.StatusOK {
+				err = output.PrintBody(cmd, resBody)
+				if err != nil {
+					return err
+				}
+
+			}
+
+			return nil
 		},
 	}
 }
